+++ src/components/mod.rs
@@ -1,7 +1,9 @@
 mod b_dropdown;
+mod b_menu;
 mod b_modal;
 mod b_navbar;
 
 pub use b_dropdown::*;
+pub use b_menu::*;
 pub use b_modal::*;
 pub use b_navbar::*;